+++ stroom-core/src/main/java/stroom/core/query/SuggestionsService.java
@@ -1,0 +1,10 @@
+package stroom.core.query;
+
+import stroom.query.shared.FetchSuggestionsRequest;
+
+import java.util.List;
+
+public interface SuggestionsService {
+
+    List<String> fetch(final FetchSuggestionsRequest request);
+}